requirements.txt
Pillow==9.3.0
pilmoji==2.0.2
